refactor: Propensity를 테이블로 저장하지 않고 Enum으로 사용하도록 수정 - #8 Propensity의 경우 수정 가능성이 매우 낮고 DB 조회가 잦을 수 있기 때문에 db 커넥션을 굳이 맺지 않고 내부적으로 관리하는게 좋다고 판단함. 추가적으로 Propensity는 연관관계없이 그냥 where문에서 동등조건으로 사용할 수 있으므로 Enum이 더 좋다고 판단.
